feat: 예약을 로그인 사용자만 할 수 있게 변경 및 예약취소는 본인 것만 할 수 있게 변경

+++ src/main/java/nextstep/reservations/domain/service/reservation/ReservationService.java
@@ -1,10 +1,14 @@
 package nextstep.reservations.domain.service.reservation;
 
+import nextstep.reservations.domain.entity.member.LoginMember;
 import nextstep.reservations.domain.entity.reservation.Reservation;
 import nextstep.reservations.dto.reservation.ReservationRequestDto;
 import nextstep.reservations.dto.reservation.ReservationResponseDto;
+import nextstep.reservations.dto.reservation.TimeTable;
+import nextstep.reservations.exceptions.auth.exception.AuthorizationException;
 import nextstep.reservations.exceptions.reservation.exception.DuplicateReservationException;
 import nextstep.reservations.exceptions.reservation.exception.NoSuchReservationException;
+import nextstep.reservations.exceptions.reservation.exception.NotAvailableTimeException;
 import nextstep.reservations.exceptions.theme.exception.NoSuchThemeException;
 import nextstep.reservations.repository.reservation.ReservationRepository;
 import nextstep.reservations.util.mapper.ReservationMapper;
@@ -23,9 +27,12 @@ public class ReservationService {
         this.reservationMapper = reservationMapper;
     }
 
-    public Long addReservation(final ReservationRequestDto requestDto) {
+    public Long addReservation(final LoginMember loginMember, final ReservationRequestDto requestDto) {
+        if (!TimeTable.values.contains(requestDto.getTime())) {
+            throw new NotAvailableTimeException();
+        }
         try {
-            return reservationRepository.add(reservationMapper.requestDtoToReservation(requestDto));
+            return reservationRepository.add(reservationMapper.requestDtoToReservation(loginMember, requestDto));
         }
         catch (DuplicateKeyException e) {
             throw new DuplicateReservationException();
@@ -47,7 +54,17 @@ public class ReservationService {
         return reservationMapper.reservationToResponseDto(reservation);
     }
 
-    public void deleteReservation(final Long id) {
+    public void deleteReservation(final LoginMember loginMember, final Long id) {
+        Reservation reservation;
+        try {
+            reservation = reservationRepository.findById(id);
+        }
+        catch (EmptyResultDataAccessException e) {
+            throw new NoSuchReservationException();
+        }
+        if (!reservation.getMember().getId().equals(loginMember.getId())) {
+            throw new AuthorizationException();
+        }
         int removeCount = reservationRepository.remove(id);
 
         if (removeCount == 0) throw new NoSuchReservationException();
